+++ packages/vsce/src/doc/resources/IJVMServer.ts
@@ -2,5 +2,5 @@ import { IResource } from "./IResource";
 
 export interface IJVMServer extends IResource {
     name: string;
-    enableStatus: string;
+    enablestatus: string;
 }