+++ packages/sdk/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the IBM® CICS® Plug-in for Zowe CLI will be documented in this file.
 
-## Recent Changes
+## `6.12.0`
 
 - Added headers as optional parameter to getResource and getCache methods.